Changed directory of the file pages '*_page.dart'
Placed them into a dedicated pages folder as to tidy up the files.

Since removing the navigator_service class which contained the navigation between pages. Went back to fix the navigations using the routes done in the 'main.dart' file.

Updated the color of the text inside the textfields as to make it more representable of it's function.